Clean package.
