--- doc/footer.html
@@ -1,10 +1,0 @@
-<br><br>
-<hr>
-<center>
-  <img src="$relpath$pictures/footer.jpg" height="100" alt="footer">
-  <br>hpp-corbaserver library documentation</br>
-</center>
-<hr>
-</center>
-</body>
-</html>
