--- doc/header.html
@@ -1,6 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
-<html><head><meta http-equiv="Content-Type" content="text/html;charset=UTF-8">
-<title>$title</title>
-<link href="$relpath$tabs.css" rel="stylesheet" type="text/css">
-<link href="$relpath$doxygen.css" rel="stylesheet" type="text/css">
-</head><body>
