--- doc/pictures/archi.fig
@@ -1,81 +1,0 @@
-#FIG 3.2
-Portrait
-Center
-Metric
-Letter  
-100.00
-Single
--2
-1200 2
-6 7105 319 10699 967
-3 2 1 1 0 7 50 -1 -1 4.000 0 1 0 3
-	3 1 1.00 56.20 112.40
-	 7105 658 8135 396 9128 967
-	 0.000 -1.000 0.000
-4 0 0 50 -1 0 12 0.0000 4 180 1740 8959 574 dynamic library loading\001
--6
-6 7535 4772 9456 5019
-2 1 1 1 0 7 50 -1 -1 4.000 0 0 -1 1 0 2
-	3 1 1.00 56.20 112.40
-	 7535 4991 9456 4991
-4 0 0 50 -1 0 12 0.0000 4 180 1815 7638 4907 compilation dependence\001
--6
-2 1 0 2 0 7 50 -1 -1 0.000 0 0 -1 0 0 2
-	 1351 1818 2192 1818
-2 1 0 2 0 7 50 -1 -1 0.000 0 0 -1 0 0 2
-	 1351 2064 2192 2064
-2 1 0 2 0 7 50 -1 -1 0.000 0 0 -1 0 0 3
-	 2106 1724 2368 1923 2131 2136
-2 2 0 2 0 7 50 -1 -1 0.000 0 0 7 0 0 5
-	 2397 4277 4293 4277 4293 4951 2397 4951 2397 4277
-2 2 0 2 0 7 50 -1 -1 0.000 0 0 7 0 0 5
-	 2397 3176 4293 3176 4293 3851 2397 3851 2397 3176
-2 1 0 1 0 7 50 -1 -1 0.000 0 0 -1 0 0 2
-	 2385 1687 4296 1687
-2 1 0 1 0 7 50 -1 -1 0.000 0 0 -1 0 0 2
-	 3294 1687 3294 2493
-2 1 0 1 0 7 50 -1 -1 0.000 0 0 -1 0 0 2
-	 2385 1950 3294 1950
-2 1 0 1 0 7 50 -1 -1 0.000 0 0 -1 0 0 2
-	 2376 2221 3284 2221
-2 2 0 2 0 7 50 -1 -1 0.000 0 0 -1 0 0 5
-	 5291 2493 7211 2493 7211 1369 5291 1369 5291 2493
-2 1 0 1 0 7 50 -1 -1 4.000 0 0 -1 1 0 3
-	3 1 1.00 56.20 112.40
-	 4293 3497 6213 3497 6213 2504
-2 1 1 1 0 7 50 -1 -1 4.000 0 0 -1 1 0 2
-	3 1 1.00 56.20 112.40
-	 3343 2504 3343 3178
-2 1 1 1 0 7 50 -1 -1 4.000 0 0 -1 1 0 2
-	3 1 1.00 56.20 112.40
-	 3343 3843 3343 4265
-2 2 0 2 0 7 50 -1 -1 0.000 0 0 -1 0 0 5
-	 2385 2493 4306 2493 4306 1369 2385 1369 2385 2493
-2 2 1 1 0 7 50 -1 -1 4.000 0 0 -1 0 0 5
-	 1998 949 4734 949 4734 5052 1998 5052 1998 949
-2 2 0 2 0 7 50 -1 -1 0.000 0 0 -1 0 0 5
-	 8826 2493 10746 2493 10746 1369 8826 1369 8826 2493
-2 1 0 1 0 7 50 -1 -1 0.000 0 0 -1 1 0 2
-	3 1 1.00 60.00 120.00
-	 7212 1913 8824 1913
-2 2 1 1 0 7 50 -1 -1 4.000 0 0 -1 0 0 5
-	 1835 769 7355 769 7355 5159 1835 5159 1835 769
-2 1 1 1 0 7 50 -1 -1 4.000 0 0 -1 1 0 2
-	3 1 1.00 60.00 120.00
-	 5295 1919 4305 1919
-2 1 1 1 0 7 50 -1 -1 4.000 0 0 -1 0 1 3
-	3 1 1.00 60.00 120.00
-	 4295 4609 9885 4609 9885 2499
-4 1 0 50 -1 0 14 0.0000 4 150 1080 3324 4694 KineoWorks\001
-4 0 0 50 -1 0 12 0.0000 4 90 480 2507 1865 server\001
-4 0 0 50 -1 0 12 0.0000 4 135 420 2507 2399 client\001
-4 0 0 50 -1 0 12 0.0000 4 180 735 2507 2146 clientHpp\001
-4 0 0 50 -1 0 14 0.0000 4 150 1215 2092 1202 Corba module\001
-4 1 0 50 -1 0 14 0.0000 4 195 1200 6162 1622 Kpp Interface\001
-4 1 0 50 -1 0 14 0.0000 4 150 1380 3256 1622 Corba Interface\001
-4 1 0 50 -1 0 14 0.0000 4 195 990 3294 3579 HPP library\001
-4 1 0 50 -1 0 14 0.0000 4 150 1560 9776 1622 KineoPathPlanner\001
-4 0 0 50 -1 0 12 0.0000 4 135 1215 7476 1793 initializeModule\001
-4 0 0 50 -1 0 12 0.0000 4 135 450 1376 1997 Corba\001
-4 0 0 50 -1 0 14 0.0000 4 150 1065 5012 1092 KPP module\001
-4 0 0 50 -1 0 12 0.0000 4 135 1350 4806 3406 send notifications\001
