+++ src/server.cc
@@ -13,7 +13,7 @@
 #include <iostream>
 #include <stdexcept>
 
-#include "gepetto/viewer/corba/server.hh"
+#include "server.hh"
 #include "server-private.hh"
 
 namespace gepetto {
@@ -25,20 +25,6 @@ namespace gepetto {
     using CORBA::ORB_init;
     using CORBA::PolicyList;
     using omniORB::fatalException;
-
-    //namespace
-    //{
-      /// \brief Forward logging messages to hpp logging mechanism.
-      /// If debug is disabled, CORBA logging will be disabled too.
-      ///
-      /// Tracing has to be enabled in your ``omniORB.cfg'' to use this
-      /// feature.
-      /// See ``omniORB configuration and API'' > ``Tracing options''
-      /// section of omniORB manual for more information.
-      //void logFunction (const char* msg);
-
-    //} // end of anonymous namespace.
-
 
     Server::Server(WindowsManagerPtr_t wm, int argc, const char *argv[],
         bool inMultiThread, bool useNameService)
@@ -54,7 +40,6 @@ namespace gepetto {
       private_->qparent (parent);
     }
 
-    /// \brief Shutdown CORBA server
     Server::~Server()
     {
       private_->deactivateAndDestroyServers();
@@ -147,7 +132,6 @@ namespace gepetto {
       pman->activate();
     }
 
-    /// \brief If CORBA requests are pending, process them
     int Server::processRequest (bool loop)
     {
       if (loop)
